+++ tests/test_input_version.py
@@ -1,44 +1,8 @@
 from copy import deepcopy
-from importlib.resources import files
 
 import pytest
-import yaml
-from packaging.version import Version
-from yaml import Loader, Dumper
 
 from stormworkflow.main import handle_input_version, CUR_INPUT_VER
-
-
-refs = files('tests.data.refs')
-input_v0_0_1 = refs.joinpath('input_v0.0.1.yaml')
-input_v0_0_2 = refs.joinpath('input_v0.0.2.yaml')
-input_v0_0_3 = refs.joinpath('input_v0.0.3.yaml')
-
-
-def read_conf(infile):
-    with open(infile, 'r') as yfile:
-        conf = yaml.load(yfile, Loader=Loader)
-    return conf
-
-
-@pytest.fixture
-def conf_v0_0_1():
-    return read_conf(input_v0_0_1)
-
-
-@pytest.fixture
-def conf_v0_0_2():
-    return read_conf(input_v0_0_2)
-
-
-@pytest.fixture
-def conf_v0_0_3():
-    return read_conf(input_v0_0_3)
-
-
-@pytest.fixture
-def conf_latest(conf_v0_0_3):
-    return conf_v0_0_3
 
 
 def test_no_version_specified(conf_latest):
@@ -76,3 +40,8 @@ def test_v0_0_1_to_latest(conf_v0_0_1, conf_latest):
 def test_v0_0_2_to_latest(conf_v0_0_2, conf_latest):
     handle_input_version(conf_v0_0_2)
     assert conf_latest == conf_v0_0_2
+
+
+def test_v0_0_3_to_latest(conf_v0_0_3, conf_latest):
+    handle_input_version(conf_v0_0_3)
+    assert conf_latest == conf_v0_0_3
